feat(): adds the "inline" temlate helper
One can use "inline" to render the response body of an HTTP request in a
specific place of the template.

For example:
<script type="text/javascript">{{ inline "https://code.jquery.com/jquery-2.2.2.min.js" }}</script>

In addition, all files in the root directory of the theme are now parsed
and can now be used as partial content, using the built-in go "template" helper.

For example:
<script type="text/javascript">{{ template "my-slider.js" }}</script>