feat(): introduces the 'themes' commands set to ease themes management
--- DOWNLOAD THEMES ----

`postmanerator themes get default` will download the theme 'default'
and copy it the $HOME/.postmanerator/themes/default folder.

Another name than 'default' can be provided, and Postmanerator will
download it as well if it is indexed in the official themes repository.
This repository is located on Github and makes use of git submodules to
index the themes. Downloading a theme actually performs a `git clone`
command, so git must be installed in order to use `themes get`.

The local themes directory can be changed by setting the
POSTMANERATOR_PATH environnement variable.

Also, one can provide a full repository URL, as in the following
example:

`postmanerator themes get \
    https://github.com/aubm/postmanerator-markdown-theme.git`

The local name of the theme can be changed using the -local-name flag.

`postmanerator -local-name="markdown" themes get \
    https://github.com/aubm/postmanerator-markdown-theme.git`

--- DELETE THEMES ---

`postmanerator themes delete default` will delete the local 'default'
theme.

--- LIST THEMES ---

`postmanerator themes list` will print the list of local themes in the
stdout.

--- DOCUMENTATION GENERATION ---

As templates.ParseGlob fails when a directory is found in the scope,
changes have been made in order to use the templates.ParseFiles func
instead.

This has led to the ability to use only the theme name as a value for
the -theme flag, in addition to the ability to provide a full folder
path.

Close #8