feat: simplifies the postman.Collection API
Now the structure used for parsing the JSON collection file is not part
of the postman package API.
Instead, it is used to build a variable of type Collection which is part
of the postman package API.
The Collection struct has a much cleaner API, which eliminates the need
for the `findRequest` template helper.

Doing so creates an abstraction around the Postman collection format
that will facilitate the addition of the support for the collection V2
format.

This commit contains breaking changes in the template API, so the
`default` and `markdown` themes will need some updates.

Related to #17